+++ src/modules/increment_mod.f90
@@ -32,6 +32,7 @@ module increment_mod
   integer           :: writeNumBits          ! number of bits to use when writing analysis and high-res increment
   logical           :: writeHiresIncrement   ! choose to write the high-res increment to a file
   logical           :: imposeRttovHuLimits   ! choose to impose "rttov" humidity limits to analysis
+  logical           :: imposeQcLimits        ! choose to impose QC limits to analysis
   logical           :: useAnalIncMask        ! for LAM only, choose to apply scale factor from a mask file to the increment  
   character(len=12) :: etiket_anlm           ! 'etiket' used when writing the analysis
   character(len=12) :: etiket_rehm           ! 'etiket' used when writing the high-res increment
@@ -58,7 +59,7 @@ CONTAINS
     logical, save :: nmlAlreadyRead = .false.
 
     NAMELIST /NAMINC/ writeHiresIncrement, etiket_rehm, etiket_anlm, &
-         etiket_rebm, writeNumBits, imposeRttovHuLimits, hInterpolationDegree, &
+         etiket_rebm, writeNumBits, imposeRttovHuLimits, imposeQcLimits, hInterpolationDegree, &
          useAnalIncMask, applyLiebmann, SSTSpread, SSTSpreadMaxBoxSize, SSTSubgrid
 
     if ( .not. nmlAlreadyRead ) then
@@ -67,7 +68,8 @@ CONTAINS
       !- Setting default values
       writeHiresIncrement = .true.
       imposeRttovHuLimits = .true.
-      useAnalIncMask      = .false.
+      imposeQcLimits = .false.
+      useAnalIncMask = .false.
       etiket_rehm = 'INCREMENT'
       etiket_rebm = 'INCREMENT'
       etiket_anlm = 'ANALYSIS'
@@ -453,6 +455,9 @@ CONTAINS
     call msg('inc_analPostProcessing', 'calling qlim_saturationLimit')
     call qlim_saturationLimit(stateVectorAnal)
     if (imposeRttovHuLimits) call qlim_rttovLimit(stateVectorAnal)
+
+    ! Impose limits on QC analysis
+    if (imposeQcLimits) call qlim_applyQcLimit(stateVectorAnal)
 
     if (gsv_varKindExist('CH')) then
       ! Apply boundaries to analysis of CH kind variables as needed.